feat(model): add id validation in delete method

+++ src/models/MongoModel.ts
@@ -32,6 +32,7 @@ abstract class MongoModel<T> implements IModel<T> {
   }
 
   public async delete(_id: string): Promise<T | null> {
+    if (!isValidObjectId(_id)) throw new Error(ErrorTypes.InvalidId);
     return this._model.findByIdAndDelete(_id);
   }
 }
